+++ app/(authorization)/blog/page.jsx
