+++ app/(authorization)/layout.jsx
@@ -1,5 +1,5 @@
-import { Footer } from "../components/headerFooterContent/Footer";
-import Header from "../components/headerFooterContent/Header";
+import { Footer } from "../../components/headerFooterContent/Footer";
+import Header from "../../components/headerFooterContent/Header";
 // import { cookies } from "next/headers";
 // import { AUTH_COOKIE_KEY } from "../contants";
 // import { redirect } from "next/navigation";
