+++ app/(authorization)/product/[id]/page.jsx
