+++ app/(authorization)/singleBlog/[id]/page.jsx
