+++ middleware.js
@@ -3,19 +3,21 @@ import { AUTH_COOKIE_KEY } from "./app/contants";
 import { cookies } from "next/headers";
 
 export function middleware(request) {
-  const cookieStore = cookies().get(AUTH_COOKIE_KEY);
+  const cookieStore = cookies();
+  const cookie = cookieStore.get(AUTH_COOKIE_KEY);
   const { pathname } = request.nextUrl;
+  console.log("tea");
 
-  if (!cookieStore?.value && !pathname.startsWith("/logIn")) {
+  if (!cookie?.value && !pathname.startsWith("/logIn")) {
     return NextResponse.redirect(new URL("/logIn", request.url));
   }
 
-  if (cookieStore?.value && pathname.startsWith("/logIn")) {
+  if (cookie?.value && pathname.startsWith("/logIn")) {
     return NextResponse.redirect(new URL("/", request.url));
   }
   // return NextResponse.next();
 }
 
 export const config = {
-  matcher: ["/((?!api|_next/static|_next/image|images|favicon.ico).*)"],
+  matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
 };
